+++ src/lib.rs
@@ -69,6 +69,8 @@ pub struct TransformConfig {
     pub background: String,
     /// Random seed
     pub seed: u64,
+    /// Maximum loop iterations
+    pub loop_limit: u32,
     /// Add source metadata to output
     pub add_metadata: bool,
 }
@@ -82,6 +84,7 @@ impl Default for TransformConfig {
             add_auto_defs: true,
             background: "none".to_owned(),
             seed: 0,
+            loop_limit: 1000,
             add_metadata: false,
         }
     }